# docs/openapi/swagger.yaml
openapi: "3.0.3"
info:
  title: "STAIR Platform API"
  description: "API for stair configuration, pricing, and manufacturing"
  version: "1.0.0"
servers:
  - url: "http://localhost:8080/api/v1"
    description: "Local development"
  - url: "https://api.stairplatform.com/api/v1"
    description: "Production"
...
paths:
  /admin/analytics/cost:
    get:
      tags: ["admin"]
      summary: "Получить cost"
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: object
        401:
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        429:
          description: "Rate limit exceeded"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        400:
          description: "Invalid request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: "Not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        422:
          description: "Validation error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: "Internal error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/analytics/manufacturing:
    get:
      tags: ["admin"]
      summary: "Получить manufacturing"
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: object
        401:
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        429:
          description: "Rate limit exceeded"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        400:
          description: "Invalid request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: "Not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        422:
          description: "Validation error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: "Internal error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/analytics/projects:
    get:
      tags: ["admin"]
      summary: "Получить projects"
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: object
        401:
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        429:
          description: "Rate limit exceeded"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        400:
          description: "Invalid request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: "Not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        422:
          description: "Validation error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: "Internal error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/analytics/usage:
    get:
      tags: ["admin"]
      summary: "Получить usage"
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: object
        401:
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        429:
          description: "Rate limit exceeded"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        400:
          description: "Invalid request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: "Not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        422:
          description: "Validation error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: "Internal error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/api-keys:
    get:
      tags: ["admin"]
      summary: "Получить api-keys"
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: object
        401:
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        429:
          description: "Rate limit exceeded"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        400:
          description: "Invalid request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: "Not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        422:
          description: "Validation error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: "Internal error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags: ["admin"]
      summary: "Создать api-keys"
      security:
        - bearerAuth: []
        - cookieAuth: []
        - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        201:
          description: "OK"
          content:
            application/json:
              schema:
                type: object
        401:
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: "Forbidden (в т.ч. CSRF-токен)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        429:
          description: "Rate limit exceeded"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        400:
          description: "Invalid request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: "Not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        422:
          description: "Validation error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: "Internal error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/api-keys/{id}:
    delete:
      tags: ["admin"]
      summary: "Удалить {id}"
      security:
        - bearerAuth: []
        - cookieAuth: []
        - csrfToken: []
      parameters:
        - name: id
          in: path
          required: true
          description: "Идентификатор id"
          schema:
            type: string
      responses:
        200:
          description: "Ключ отозван"
          content:
            application/json:
              schema:
                type: object
        401:
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: "Forbidden (в т.ч. CSRF-токен)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        429:
          description: "Rate limit exceeded"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        400:
          description: "Invalid request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: "Not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        422:
          description: "Validation error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: "Internal error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/export:
    get:
      tags: ["admin"]
      summary: "Выгрузка данных (CSV/JSON)"
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: scope
          in: query
          required: true
          description: "Область выгрузки"
          schema:
            type: string
            enum: [users, projects, audit]
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: object
        401:
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        429:
          description: "Rate limit exceeded"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        400:
          description: "Invalid request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: "Not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        422:
          description: "Validation error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: "Internal error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /auth/login:
    post:
      tags: ["auth"]
      summary: "Войти в систему (сессионная cookie)"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        200:
          description: "Сессия создана, установлена session cookie"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        400:
          description: "Invalid request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: "Not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        422:
          description: "Validation error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        500:
          description: "Internal error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: "bearer"
      description: "API-ключ (Authorization: Bearer) — EDR-0016"
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: "Сессионная cookie (httpOnly, SameSite=Strict); для admin — session_admin"
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: "Double-submit CSRF: значение cookie csrf (csrf_admin) должно быть продублировано в этом заголовке для всех изменяющих запросов"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: "Машинный код ошибки (not_found, forbidden, validation_error, invalid_input, rates_not_allowed, csrf, internal и др.)"
            message:
              type: string
              description: Человекочитаемое сообщение
            request_id:
              type: string
              description: "Идентификатор запроса (только для 5xx)"
        request_id:
          type: string
          description: "Идентификатор запроса для поддержки"
...
    Session:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
          description: "Сессионный токен. Дублирует httpOnly cookie session (session_admin) и нужен только небраузерным клиентам; CSRF-токен берётся из cookie csrf и дублируется в заголовке X-CSRF-Token"
...