# internal/transport/http/router.go
package http

import (
	"context"
	"encoding/json"
	"log/slog"
	"net/http"
	"time"
)

// NewRouter собирает маршруты API v1. svc — прикладной сервис расчёта,
// projects — прикладной сервис управления проектами, authSvc — сервис
// аутентификации (инверсия зависимостей, DOM-0008); nil недопустимы.
// Все /api/v1-маршруты, кроме регистрации и входа, требуют аутентификации
// (SEC-0003); мутирующие — CSRF (double-submit). Роутер оборачивает все
// маршруты middleware логгирования/request-id (наблюдаемость).
func NewRouter(svc StairService, projects ProjectService, authSvc AuthService, cfg Config, auditSvc ...AuditService) http.Handler {
	// P2-11: все ранее глобальные переключатели (rate-limit, cookie, секреты,
	// region) живут как локальные значения инстанса роутера — никаких
	// applyConfig-мутаций пакетного состояния.
	if cfg.LoginRateLimit <= 0 {
		cfg.LoginRateLimit = DefaultConfig().LoginRateLimit
	}
	if cfg.LoginRateWindow <= 0 {
		cfg.LoginRateWindow = DefaultConfig().LoginRateWindow
	}
	if cfg.RegisterRateLimit <= 0 {
		cfg.RegisterRateLimit = DefaultConfig().RegisterRateLimit
	}
	if cfg.RegisterRateWindow <= 0 {
		cfg.RegisterRateWindow = DefaultConfig().RegisterRateWindow
	}
	if cfg.QuoteRateLimit <= 0 {
		cfg.QuoteRateLimit = DefaultConfig().QuoteRateLimit
	}
	if cfg.QuoteRateWindow <= 0 {
		cfg.QuoteRateWindow = DefaultConfig().QuoteRateWindow
	}
	if cfg.ValidateRateLimit <= 0 {
		cfg.ValidateRateLimit = DefaultConfig().ValidateRateLimit
	}
	if cfg.ValidateRateWindow <= 0 {
		cfg.ValidateRateWindow = DefaultConfig().ValidateRateWindow
	}
	if cfg.AuthRateLimit <= 0 {
		cfg.AuthRateLimit = DefaultConfig().AuthRateLimit
	}
	if cfg.AuthRateWindow <= 0 {
		cfg.AuthRateWindow = DefaultConfig().AuthRateWindow
	}
	if cfg.SsoRateLimit <= 0 {
		cfg.SsoRateLimit = DefaultConfig().SsoRateLimit
	}
	if cfg.SsoRateWindow <= 0 {
		cfg.SsoRateWindow = DefaultConfig().SsoRateWindow
	}
	if cfg.MaxBodyBytes <= 0 {
		cfg.MaxBodyBytes = DefaultConfig().MaxBodyBytes
	}
	SetMetricsConfig(cfg.Region, cfg.InstanceID)

	loginLimiter := newRateLimiterStrategy(context.Background(), cfg.RedisAddr, cfg.LoginRateLimit, cfg.LoginRateWindow)
	registerLimiter := newRateLimiterStrategy(context.Background(), cfg.RedisAddr, cfg.RegisterRateLimit, cfg.RegisterRateWindow)
	quoteLimiter := newRateLimiterStrategy(context.Background(), cfg.RedisAddr, cfg.QuoteRateLimit, cfg.QuoteRateWindow)
	validateLimiter := newRateLimiterStrategy(context.Background(), cfg.RedisAddr, cfg.ValidateRateLimit, cfg.ValidateRateWindow)
	// Authenticated rate limiter: 200 req/min per user/API key
	authRateLimiter := newRateLimiterStrategy(context.Background(), cfg.RedisAddr, cfg.AuthRateLimit, cfg.AuthRateWindow)
	// SSO rate limiter: публичные begin/callback/config (S-109).
	ssoLimiter := newRateLimiterStrategy(context.Background(), cfg.RedisAddr, cfg.SsoRateLimit, cfg.SsoRateWindow)
	secure := cfg.CookieSecure
	// S-120: доверенные прокси для per-IP rate-limit (XFF только от них).
	trusted := parseTrustedProxies(cfg.TrustedProxies)

	var auditsvc AuditService
	if len(auditSvc) > 0 {
		auditsvc = auditSvc[0]
	}
	integrations := cfg.Integrations
	readiness := cfg.Readiness
	storage := cfg.Storage
	payments := cfg.Payments
	paymentAdmin := cfg.PaymentAdmin
	analytics := cfg.Analytics
	jobsSvc := cfg.Jobs
	assistantSvc := cfg.Assistant
	ordersSvc := cfg.Orders

	// Response cache для read-heavy GET-запросов (5min TTL, 512 entries)
	respCacheMW, storeCacheInvalidator := NewResponseCacheWithInvalidation(ResponseCacheConfig{
		MaxEntries: 512,
		DefaultTTL: 5 * time.Minute,
		Methods:    []string{http.MethodGet},
	})

	mux := http.NewServeMux()
	mux.HandleFunc("GET /health", handleHealth(cfg.Region))
	mux.Handle("GET /metrics", InternalOnlyMiddleware(http.HandlerFunc(handleMetrics)))
	if readiness != nil {
		mux.HandleFunc("GET /ready", handleReady(readiness))
	}
	// Версионированные статические ассеты (PBR-текстуры, HDRI — этап 1).
	// Публичные и иммутабельные: файлы версионируются по содержимому.
	if cfg.StaticAssetsDir != "" {
		// Литерал, не конкатенация: регэксп swagger-sync вытаскивает пути из
		// исходника. Префикс должен совпадать с staticAssetsMount.
		mux.Handle("GET /static-assets/", StaticAssetsHandler(cfg.StaticAssetsDir))
	}

	// Swagger UI (internal only)
	mux.Handle("GET /swagger", InternalOnlyMiddleware(http.HandlerFunc(handleSwaggerUI)))
	mux.Handle("GET /docs/openapi/swagger.yaml", InternalOnlyMiddleware(http.HandlerFunc(handleSwaggerSpec)))
	mux.Handle("POST /api/v1/auth/register", limitRate(registerLimiter, trusted, handleRegister(authSvc, secure)))
	mux.Handle("POST /api/v1/auth/login", limitRate(loginLimiter, trusted, handleLogin(authSvc, secure)))

	// Публичный расчёт предварительной цены для клиентского сайта (store).
	// Без аутентификации; rate-limiter защищает от злоупотреблений.
	// Прайс платных услуг для витрины (этап 4): суммы — только из серверного
	// каталога тарифов (S-150), клиент на них не влияет.
	if payments != nil {
		mux.Handle("GET /api/v1/public/payment-tiers", handlePublicPaymentTiers(payments))
	}

	// Настройки магазина для витрины (волна 0): контакты, реквизиты, соцсети,
	// SEO и коды счётчиков. Прайс материалов отдаётся каталогом materials.
	storeSvc := cfg.Store
	if storeSvc != nil {
		mux.Handle("GET /api/v1/public/store-settings", limitRate(validateLimiter, trusted, handlePublicStoreSettings(storeSvc, authSvc)))
	}

	// Каталог материалов для витрины (этап 3): единственный источник истины
	// для кодов, плотностей, диапазонов и ставок — бэкенд, не фронт.
	mux.Handle("GET /api/v1/public/materials", limitRate(validateLimiter, trusted, handlePublicMaterials(storeSvc, authSvc)))
	mux.Handle("POST /api/v1/public/stairs:quote", limitRate(quoteLimiter, trusted, handlePublicQuote(svc, storeSvc, authSvc)))
	// Живая валидация при вводе для клиентского сайта (S-P5): тот же блок
	// validation, что и в расчёте, но без геометрии/производства/цены.
	mux.Handle("POST /api/v1/public/stairs:validate", limitRate(validateLimiter, trusted, handleValidate(svc)))
	// Публичная консультация (store): анонимный запрос обратной связи.
	// Создаёт заказ-лид kind=consultation без пользователя и цены.
	mux.Handle("POST /api/v1/public/orders", limitRate(quoteLimiter, trusted, handleCreateConsultation(ordersSvc, authSvc)))

	authProtected := func(next http.Handler) http.Handler {
		return requireAuth(authSvc, authRateLimiter, secure)(next)
	}
	authMutating := func(next http.Handler) http.Handler {
		return requireAuth(authSvc, authRateLimiter, secure)(requireCSRF(cfg.CSRFAllowedOrigins, next))
	}

	mux.Handle("GET /api/v1/auth/me", authProtected(handleMe()))
	mux.Handle("POST /api/v1/auth/logout", authMutating(handleLogout(authSvc, secure)))

	// SSO (EDR-0017 §6): публичные маршруты (начала и колбэк). Rate-limited
	// по IP (S-109), чтобы исключить спам и неограниченный рост sso_states.
	mux.Handle("GET /api/v1/auth/sso", limitRate(ssoLimiter, trusted, handleSsoBegin(authSvc)))
	mux.Handle("GET /api/v1/auth/sso/callback", limitRate(ssoLimiter, trusted, handleSsoCallback(authSvc, secure)))
	mux.Handle("GET /api/v1/auth/sso/config", limitRate(ssoLimiter, trusted, handleSsoConfig(authSvc)))

	mux.Handle("GET /api/v1/admin/users", authProtected(handleListUsers(authSvc)))
	mux.Handle("PATCH /api/v1/admin/users/{id}", authMutating(handleUpdateUser(authSvc)))
	mux.Handle("GET /api/v1/admin/overview", authProtected(handleAdminOverview(authSvc, projects)))
	mux.Handle("GET /api/v1/admin/settings", authProtected(handleGetSettings(authSvc)))
	mux.Handle("PUT /api/v1/admin/settings", authMutating(handleUpdateSettings(authSvc, auditsvc)))
	mux.Handle("GET /api/v1/admin/export", authProtected(handleExport(authSvc, projects, auditsvc)))
	mux.Handle("GET /api/v1/admin/api-keys", authProtected(handleListApiKeys(authSvc)))
	mux.Handle("POST /api/v1/admin/api-keys", authMutating(handleCreateApiKey(authSvc)))
	mux.Handle("DELETE /api/v1/admin/api-keys/{id}", authMutating(handleRevokeApiKey(authSvc)))

	// Настройки и прайс магазина (волна 0 «store admin»): контакты, реквизиты,
	// соцсети, SEO, параметры расчёта и цены материалов в ₽/кг.
	if storeSvc != nil {
		mux.Handle("GET /api/v1/admin/store/settings", authProtected(handleGetStoreSettings(storeSvc)))
		mux.Handle("PUT /api/v1/admin/store/settings", authMutating(handleUpdateStoreSettings(storeSvc, storeCacheInvalidator)))
		mux.Handle("GET /api/v1/admin/store/prices", authProtected(handleListMaterialPrices(storeSvc)))
		mux.Handle("PUT /api/v1/admin/store/prices", authMutating(handleSetMaterialPrice(storeSvc, storeCacheInvalidator)))
		mux.Handle("DELETE /api/v1/admin/store/prices/{code}", authMutating(handleDeleteMaterialPrice(storeSvc, storeCacheInvalidator)))
	}

	if auditsvc != nil {
		mux.Handle("GET /api/v1/audit", authProtected(handleListTenantAudit(auditsvc)))
		mux.Handle("GET /api/v1/projects/{id}/audit", authProtected(handleListProjectAudit(projects, auditsvc)))
		// Клиентские события (клики, применение вариантов) — mutating (CSRF).
		mux.Handle("POST /api/v1/audit", authMutating(handleRecordAudit(auditsvc)))
	}

	mux.Handle("POST /api/v1/stairs:calculate", authMutating(handleCalculate(svc, auditsvc)))
	mux.Handle("POST /api/v1/stairs:validate", authMutating(handleValidate(svc)))
	mux.Handle("POST /api/v1/stairs:optimize", authMutating(handleOptimize(svc)))
	if assistantSvc != nil {
		// AI-ассистенты (Phase D, D1–D4): design/engineering/manufacturing/pricing.
		// S-144 (S-141 №7): мутирующие POST (LLM-бюджет, запись conversation-memory)
		// — CSRF (double-submit), как у остальных mutating-роутов.
		mux.Handle("POST /api/v1/assistant/{kind}", authMutating(handleAssistantAsk(assistantSvc)))
		// Право на забвение conversation-memory (S-141 №13): purge памяти
		// проекта, членство проверяет прикладной слой (S-142 → 403).
		mux.Handle("DELETE /api/v1/assistant/memory", authMutating(handleAssistantForget(assistantSvc)))
	}
	if ordersSvc != nil {
		// Розничные заказы (клиентский сайт, Store). Клиентский кабинет и
		// админ-раздел для менеджера.
		mux.Handle("POST /api/v1/orders", authMutating(handleCreateOrder(ordersSvc)))
		mux.Handle("GET /api/v1/orders", authProtected(handleListMyOrders(ordersSvc)))
		mux.Handle("GET /api/v1/admin/orders", authProtected(handleAdminListOrders(ordersSvc)))
		mux.Handle("PATCH /api/v1/admin/orders/{id}/status", authMutating(handleAdminUpdateOrderStatus(ordersSvc, auditsvc)))
	}
	if testimonials := cfg.Testimonials; testimonials != nil {
		// Отзывы клиентов (Store): публичные — для лендинга, admin — для менеджера.
		mux.Handle("GET /api/v1/public/testimonials", handlePublicListTestimonials(testimonials, authSvc))
		mux.Handle("GET /api/v1/admin/testimonials", authProtected(handleAdminListTestimonials(testimonials)))
		mux.Handle("POST /api/v1/admin/testimonials", authMutating(handleAdminCreateTestimonial(testimonials, auditsvc)))
		mux.Handle("PATCH /api/v1/admin/testimonials/{id}", authMutating(handleAdminUpdateTestimonial(testimonials, auditsvc)))
		mux.Handle("DELETE /api/v1/admin/testimonials/{id}", authMutating(handleAdminDeleteTestimonial(testimonials, auditsvc)))
	}
	if jobsSvc != nil {
		mux.Handle("POST /api/v1/stairs:calculate/async", authMutating(handleCalculateAsync(jobsSvc)))
		mux.Handle("GET /api/v1/jobs/{id}", authProtected(handleGetJob(jobsSvc)))
	}
	mux.Handle("POST /api/v1/projects", authMutating(handleCreateProject(projects)))
	mux.Handle("GET /api/v1/projects", authProtected(handleListProjects(projects)))
	mux.Handle("GET /api/v1/projects/{id}", authProtected(handleGetProject(projects)))
	mux.Handle("GET /api/v1/projects/{id}/members", authProtected(handleListMembers(projects)))
	mux.Handle("POST /api/v1/projects/{id}/members", authMutating(handleAddMember(projects)))
	mux.Handle("PATCH /api/v1/projects/{id}/members/{userID}", authMutating(handleUpdateMemberRole(projects)))
	mux.Handle("DELETE /api/v1/projects/{id}/members/{userID}", authMutating(handleRemoveMember(projects)))
	mux.Handle("GET /api/v1/projects/{id}/comments", authProtected(handleListComments(projects)))
	mux.Handle("POST /api/v1/projects/{id}/comments", authMutating(handleAddComment(projects)))
	mux.Handle("DELETE /api/v1/projects/{id}/comments/{commentID}", authMutating(handleDeleteComment(projects)))
	mux.Handle("GET /api/v1/projects/{id}/reviews", authProtected(handleListReviews(projects)))
	mux.Handle("POST /api/v1/projects/{id}/review", authMutating(handleRequestReview(projects)))
	mux.Handle("POST /api/v1/projects/{id}/reviews/{reviewID}/sign-off", authMutating(handleSignOffReview(projects)))
	mux.Handle("POST /api/v1/projects/{id}/reviews/{reviewID}/changes", authMutating(handleRequestChanges(projects)))
	mux.Handle("GET /api/v1/projects/{id}/approvals", authProtected(handleListApprovals(projects)))
	mux.Handle("POST /api/v1/projects/{id}/configurations/{configID}/approve", authMutating(handleApproveConfiguration(projects)))
	mux.Handle("GET /api/v1/projects/{id}/configurations/{configID}/approval", authProtected(handleGetConfigurationApproval(projects)))
	mux.Handle("GET /api/v1/projects/{id}/configurations", authProtected(handleListConfigurations(projects)))
	mux.Handle("GET /api/v1/projects/{id}/configurations/{configID}", authProtected(handleGetConfiguration(projects)))
	mux.Handle("POST /api/v1/projects/{id}/configurations/{configID}/restore", authMutating(handleRestoreConfiguration(projects)))
	mux.Handle("POST /api/v1/projects/{id}/calculate", authMutating(handleCalculateProject(projects)))
	mux.Handle("POST /api/v1/projects/{id}/preview", authMutating(handlePreviewProject(projects)))
	mux.Handle("POST /api/v1/projects/{id}/optimize", authMutating(handleOptimizeProject(projects)))
	mux.Handle("GET /api/v1/projects/{id}/export", authProtected(handleExportProject(projects)))
	mux.Handle("GET /api/v1/projects/{id}/export/cad", authProtected(handleExportCAD(projects)))

	if storage != nil {
		mux.Handle("POST /api/v1/projects/{id}/export/cad/store", authMutating(handleStoreExportCAD(projects, storage)))
		mux.Handle("GET /api/v1/storage/{key...}", authProtected(handleGetObject(storage)))
		mux.Handle("DELETE /api/v1/storage/{key...}", authMutating(handleDeleteObject(storage)))
	}

	if integrations != nil {
		mux.Handle("GET /api/v1/integrations/endpoints", authProtected(handleListEndpoints(integrations)))
		mux.Handle("POST /api/v1/integrations/endpoints", authMutating(handleCreateEndpoint(integrations)))
		mux.Handle("DELETE /api/v1/integrations/endpoints/{id}", authMutating(handleDeleteEndpoint(integrations)))
		mux.Handle("POST /api/v1/projects/{id}/quote-send", authMutating(handleQuoteSend(projects, integrations)))
		mux.Handle("POST /api/v1/projects/{id}/crm-sync", authMutating(handleProjectSync(projects, integrations)))
		mux.Handle("POST /api/v1/projects/{id}/order-send", authMutating(handleOrderSend(projects, integrations)))
	}

	if payments != nil {
		// Входящий webhook PSP публичный (вызывает внешняя система); публичность
		// безопасна — подпись верифицируется (EDR-0027 §3.4).
		mux.HandleFunc("POST /api/v1/payments/webhook", handlePaymentWebhook(payments, cfg.PaymentsWebhookSecret))
		mux.Handle("POST /api/v1/projects/{id}/checkout", authMutating(handleCheckout(projects, payments)))
		mux.Handle("GET /api/v1/projects/{id}/payments", authProtected(handleListPayments(projects, payments)))
		mux.Handle("GET /api/v1/payments/{id}", authProtected(handleGetPayment(payments)))
		// Этап 4: покупка услуги с витрины (auth, CSRF) и мои покупки.
		mux.Handle("POST /api/v1/public/services/checkout", authMutating(handleServiceCheckout(payments)))
		mux.Handle("GET /api/v1/payments/mine", authProtected(handleListMyPayments(payments)))
	}

	if paymentAdmin != nil {
		mux.Handle("GET /api/v1/admin/payments", authProtected(handleAdminListPayments(paymentAdmin)))
		mux.Handle("POST /api/v1/admin/payments/{id}/refund", authMutating(handleAdminRefundPayment(paymentAdmin, auditsvc)))
	}

	// Stripe webhook endpoint (публичный; Stripe-Signature верификация).
	if cfg.StripeWebhookService != nil {
		mux.HandleFunc("POST /api/v1/payments/stripe/webhook", handleStripeWebhook(cfg.StripeWebhookService))
	}

	if analytics != nil {
		mux.Handle("GET /api/v1/admin/analytics/usage", authProtected(handleUsageAnalytics(analytics)))
		mux.Handle("GET /api/v1/admin/analytics/projects", authProtected(handleProjectsAnalytics(analytics)))
		mux.Handle("GET /api/v1/admin/analytics/manufacturing", authProtected(handleManufacturingAnalytics(analytics)))
		mux.Handle("GET /api/v1/admin/analytics/cost", authProtected(handleCostAnalytics(analytics)))
	}

	// WebSocket endpoint (EDR-0038). /ws — store-namespace, /ws/admin —
	// admin-namespace (S-116): браузерный WebSocket API не умеет ставить
	// заголовок X-App-Origin, поэтому namespace определяется путём.
	if cfg.WebSocketHandler != nil {
		mux.HandleFunc("GET /ws", cfg.WebSocketHandler.HandleWebSocket)
		mux.HandleFunc("GET /ws/admin", cfg.WebSocketHandler.HandleWebSocket)
	}

	// GraphQL НЕ регистрируется в production: schema обещает 15 операций,
	// а реализованы только 5 (substring-диспатч), Subscription-рут не реализован.
	// Пакет internal/transport/graphql остаётся как протестированная библиотека
	// для будущей полноценной реализации (см. docs/09_API). Не выставлять в проде,
	// пока не реализованы все операции и валидация schema.graphql.

	mux.HandleFunc("GET /", handleNotFound)

	// Cache policies по путям.
	// S-107: /api/v1/projects/ намеренно НЕ указан — проекты всегда
	// аутентифицированы и не должны попадать в браузерный кеш (наследует
	// "/api/v1/" → no-store). Дополнительно CacheMiddleware принудительно
	// снимает кеш с Private-политик при наличии identity.
	cachePolicies := map[string]CachePolicy{
		"/api/v1/":             CacheNoCache,   // API — no cache
		"/api/v1/configs/":     CacheShort,     // конфигурации — 5 min
		"/api/v1/assortments/": CacheMedium,    // ассортимент — 1 hour
		"/api/v1/materials/":   CacheMedium,    // материалы — 1 hour
		"/api/v1/profiles/":    CacheMedium,    // профили — 1 hour
		"/api/v1/stairs/":      CacheShort,     // лестницы — 5 min
		"/api/v1/auth/":        CacheNoCache,   // авторизация — no cache
		"/api/v1/health":       CacheNoCache,   // health check — no cache
		"/static/":             CacheLong,      // статика — 1 day
		"/assets/":             CacheImmutable, // webpack assets — immutable
	}

	// Deduplication для тяжёлых операций
	dedup := DeduplicateMiddleware(DeduplicateByKey)

	// Security middleware (CORS, HSTS, CSP)
	secMiddleware := SecurityMiddleware(cfg.SecurityConfig)

	// Middleware-конвейер: строится изнутри наружу (inner → outer).
	//
	// SEC-005b (2026-09-26): порядок PanicRecoveryMiddleware и withLogging
	// был переставлен. Раньше recover стоял ВНЕ withLogging, поэтому
	// PanicRecoveryMiddleware получал запрос до того, как withLogging
	// вложил в контекст request id (middleware.go:54), и поле
	// `request_id` в теле 500 всегда было пустым:
	//   -> 500 {"error":{"code":"internal",...,"request_id":""}}
	// То есть инцидент в 03:00 невозможно было связать с его строкой лога
	// по ответу клиенту. Теперь recover innermost — он видит уже
	// обогащённый контекст, поэтому request_id в 500 всегда заполнен.
	//
	// Взаимодействие с withLogging: withLogging больше НЕ re-pаникует
	// (единственный recover — здесь), а просто логирует факт паники
	// в status/логи/метрики и отдаёт управление сюда.
	h := withLogging(PanicRecoveryMiddleware(mux))
	h = secMiddleware(h)
	h = dedup(h)
	h = respCacheMW(h)
	h = CacheMiddleware(cachePolicies)(h)
	h = BodySizeLimit(cfg.MaxBodyBytes)(h)
	h = CompressionMiddleware(h)
	h = RouteTimeoutMiddleware(DefaultAPIRouteTimeouts())(h)
	h = VersionMiddleware(h)
	return TraceMiddleware(h)
}

func handleHealth(region string) http.HandlerFunc {
	return func(w http.ResponseWriter, _ *http.Request) {
		resp := map[string]any{
			"status":  "ok",
			"service": "stair-platform-api",
			"version": "1.0.0",
		}
		if region != "" {
			resp["region"] = region
		}
		writeJSON(w, http.StatusOK, resp)
	}
}

// handleNotFound — единый 404 для неизвестного пути.
//
// API-003: здесь раньше отдавалось `{"error": "not_found"}` — строка вместо
// объекта, тогда как writeError/writeServiceError отдают
// `{"error": {"code": ..., "message": ...}}`. Один клиентский парсер получал
// два несовместимых формата ошибки, а схема OpenAPI описывала объект.
func handleNotFound(w http.ResponseWriter, r *http.Request) {
	writeErrorWithRequestID(w, r, http.StatusNotFound, "not_found",
		"Метод не найден. Проверьте путь и версию API (/api/v1).")
}

func writeJSON(w http.ResponseWriter, status int, payload any) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	if err := json.NewEncoder(w).Encode(payload); err != nil {
		// Логируем ошибку кодирования —响应 будет обрезан/повреждён.
		slog.Error("writeJSON: encode failed", "error", err)
	}
}